+++ anzeigen-frontend/src/plugins/router.ts
@@ -3,7 +3,7 @@ import { createRouter, createWebHashHistory } from "vue-router";
 
 import { ROUTES_AD, ROUTES_BOARD } from "@/Constants";
 import AdBoard from "@/views/AdBoard.vue";
-import AdDetails from "@/views/AdDetails.vue";
+import AdDetailsView from "@/views/AdDetailsView.vue";
 
 const routes = [
   {
@@ -15,7 +15,7 @@ const routes = [
   {
     path: "/ad",
     name: ROUTES_AD,
-    component: AdDetails,
+    component: AdDetailsView,
     meta: {},
   },
   { path: "/:catchAll(.*)*", redirect: "/board" }, // CatchAll route
