+++ anzeigen-frontend/src/views/AdDetailsView.vue
